+++ internal/dbsqlc/river_job.sql.go
@@ -641,6 +641,66 @@ func (q *Queries) JobRescueMany(ctx context.Context, db DBTX, arg JobRescueManyP
 	return err
 }
 
+const jobRetryImmediately = `-- name: JobRetryImmediately :one
+WITH job_to_update AS (
+  SELECT
+    id
+  FROM
+    river_job
+  WHERE
+    river_job.id = $1
+  FOR UPDATE
+),
+
+updated_job AS (
+  UPDATE river_job
+  SET
+    state = 'available'::river_job_state,
+    scheduled_at = CASE WHEN scheduled_at < now() THEN scheduled_at ELSE now() END,
+    max_attempts = CASE WHEN attempt = max_attempts THEN max_attempts + 1 ELSE max_attempts END,
+    finalized_at = NULL
+  FROM job_to_update
+  WHERE river_job.id = job_to_update.id
+    -- Do not touch running jobs:
+    AND river_job.state != 'running'::river_job_state
+    -- If the job is already available with a prior scheduled_at, leave it alone.
+    AND NOT (river_job.state = 'available'::river_job_state AND river_job.scheduled_at < now())
+  RETURNING river_job.id, river_job.args, river_job.attempt, river_job.attempted_at, river_job.attempted_by, river_job.created_at, river_job.errors, river_job.finalized_at, river_job.kind, river_job.max_attempts, river_job.metadata, river_job.priority, river_job.queue, river_job.state, river_job.scheduled_at, river_job.tags
+)
+
+SELECT id, args, attempt, attempted_at, attempted_by, created_at, errors, finalized_at, kind, max_attempts, metadata, priority, queue, state, scheduled_at, tags
+FROM river_job
+WHERE id = $1::bigint
+    AND id NOT IN (SELECT id FROM updated_job)
+UNION
+SELECT id, args, attempt, attempted_at, attempted_by, created_at, errors, finalized_at, kind, max_attempts, metadata, priority, queue, state, scheduled_at, tags
+FROM updated_job
+`
+
+func (q *Queries) JobRetryImmediately(ctx context.Context, db DBTX, id int64) (*RiverJob, error) {
+	row := db.QueryRow(ctx, jobRetryImmediately, id)
+	var i RiverJob
+	err := row.Scan(
+		&i.ID,
+		&i.Args,
+		&i.Attempt,
+		&i.AttemptedAt,
+		&i.AttemptedBy,
+		&i.CreatedAt,
+		&i.Errors,
+		&i.FinalizedAt,
+		&i.Kind,
+		&i.MaxAttempts,
+		&i.Metadata,
+		&i.Priority,
+		&i.Queue,
+		&i.State,
+		&i.ScheduledAt,
+		&i.Tags,
+	)
+	return &i, err
+}
+
 const jobSchedule = `-- name: JobSchedule :one
 WITH jobs_to_schedule AS (
   SELECT id